Modularize frontend: break App.jsx into focused modules
- Walksheds.jsx: root component (state, data loading, station selection)
- MapView.jsx: Mapbox map with layers, icons, click/drag handling
- WalkshedLayers.jsx: isochrone fill/outline/label rendering
- StationPill.jsx: expanding station marker component
- useNavigation.js: keyboard, wheel, and touch navigation hooks
- mapbox.js: isochrone API calls, geometry helpers, bounds calculation
- constants.js: shared colors, styles, map config
- walksheds.css: consolidated from App.css + index.css
