--- src/index.css
@@ -1,17 +1,0 @@
-*,
-*::before,
-*::after {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-html,
-body,
-#root {
-  width: 100%;
-  height: 100%;
-  overflow: hidden;
-  color-scheme: light;
-  background: #e8e8e8;
-}
